# perf: create a roaring64.Bitmap pool