Not sure if this is worth it, as it increases a bit the CPU usage to save some allocations:

```
goos: darwin
goarch: arm64
pkg: github.com/dunglas/skipfilter
cpu: Apple M1 Pro
                                 │   old.txt   │              new.txt               │
                                 │   sec/op    │   sec/op     vs base               │
SkipFilter_MatchAny_Read-10        1.108m ± 1%   1.205m ± 1%  +8.71% (p=0.000 n=10)
SkipFilter_MatchAny_ReadWrite-10   1.705m ± 1%   1.779m ± 1%  +4.33% (p=0.000 n=10)
geomean                            1.375m        1.464m       +6.50%

                                 │   old.txt    │               new.txt               │
                                 │     B/op     │     B/op      vs base               │
SkipFilter_MatchAny_Read-10        583.2Ki ± 0%   582.7Ki ± 0%       ~ (p=0.481 n=10)
SkipFilter_MatchAny_ReadWrite-10   694.2Ki ± 1%   683.1Ki ± 1%  -1.60% (p=0.000 n=10)
geomean                            636.3Ki        630.9Ki       -0.84%

                                 │   old.txt   │              new.txt               │
                                 │  allocs/op  │  allocs/op   vs base               │
SkipFilter_MatchAny_Read-10        2.594k ± 0%   2.617k ± 1%  +0.91% (p=0.000 n=10)
SkipFilter_MatchAny_ReadWrite-10   2.959k ± 0%   2.962k ± 1%       ~ (p=0.492 n=10)
geomean                            2.770k        2.784k       +0.49%
```